+++ rust/operator-binary/src/config/jvm.rs
@@ -1,11 +1,11 @@
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_hbase_crd::{
-    HbaseConfig, HbaseConfigFragment, HbaseRole, CONFIG_DIR_NAME, JVM_SECURITY_PROPERTIES_FILE,
+    AnyServiceConfig, HbaseCluster, HbaseRole, CONFIG_DIR_NAME, JVM_SECURITY_PROPERTIES_FILE,
     METRICS_PORT,
 };
 use stackable_operator::{
     memory::{BinaryMultiple, MemoryQuantity},
-    role_utils::{self, GenericRoleConfig, JavaCommonConfig, JvmArgumentOverrides, Role},
+    role_utils::{self, JvmArgumentOverrides},
 };
 
 const JAVA_HEAP_FACTOR: f32 = 0.8;
@@ -22,6 +22,9 @@ pub enum Error {
 
     #[snafu(display("failed to merge jvm argument overrides"))]
     MergeJvmArgumentOverrides { source: role_utils::Error },
+
+    #[snafu(display("the HBase role [{role}] is missing from spec"))]
+    MissingHbaseRole { role: String },
 }
 
 // Applies to both the servers and the CLI
@@ -43,14 +46,13 @@ pub fn construct_global_jvm_args(kerberos_enabled: bool) -> String {
     jvm_args.join(" ")
 }
 
-/// JVM arguments that specifically for the role (server), so will *not* be used e.g. by CLI tools
-fn construct_role_specific_jvm_args(
+/// Arguments that go into `HBASE_OPTS`, so *not* the heap settings (which go into `HBASE_HEAPSIZE`).
+pub fn construct_role_specific_non_heap_jvm_args(
+    hbase: &HbaseCluster,
     hbase_role: &HbaseRole,
-    role: &Role<HbaseConfigFragment, GenericRoleConfig, JavaCommonConfig>,
     role_group: &str,
     product_version: &str,
-    kerberos_enabled: bool,
-) -> Result<Vec<String>, Error> {
+) -> Result<String, Error> {
     let mut jvm_args = vec![format!(
         "-Djava.security.properties={CONFIG_DIR_NAME}/{JVM_SECURITY_PROPERTIES_FILE}"
     )];
@@ -61,35 +63,46 @@ fn construct_role_specific_jvm_args(
             format!("-javaagent:/stackable/jmx/jmx_prometheus_javaagent.jar={METRICS_PORT}:/stackable/jmx/{hbase_role}.yaml")
         );
     }
-    if kerberos_enabled {
+    if hbase.has_kerberos_enabled() {
         jvm_args.push("-Djava.security.krb5.conf=/stackable/kerberos/krb5.conf".to_owned());
     }
 
     let operator_generated = JvmArgumentOverrides::new_with_only_additions(jvm_args);
-    let merged = role
-        .get_merged_jvm_argument_overrides(role_group, &operator_generated)
-        .context(MergeJvmArgumentOverridesSnafu)?;
-    Ok(merged
+
+    let merged = match hbase_role {
+        HbaseRole::Master => hbase
+            .spec
+            .masters
+            .as_ref()
+            .context(MissingHbaseRoleSnafu {
+                role: hbase_role.to_string(),
+            })?
+            .get_merged_jvm_argument_overrides(role_group, &operator_generated)
+            .context(MergeJvmArgumentOverridesSnafu)?,
+        HbaseRole::RegionServer => hbase
+            .spec
+            .region_servers
+            .as_ref()
+            .context(MissingHbaseRoleSnafu {
+                role: hbase_role.to_string(),
+            })?
+            .get_merged_jvm_argument_overrides(role_group, &operator_generated)
+            .context(MergeJvmArgumentOverridesSnafu)?,
+        HbaseRole::RestServer => hbase
+            .spec
+            .rest_servers
+            .as_ref()
+            .context(MissingHbaseRoleSnafu {
+                role: hbase_role.to_string(),
+            })?
+            .get_merged_jvm_argument_overrides(role_group, &operator_generated)
+            .context(MergeJvmArgumentOverridesSnafu)?,
+    };
+    jvm_args = merged
         .effective_jvm_config_after_merging()
         // Sorry for the clone, that's how operator-rs is currently modelled :P
-        .clone())
-}
+        .clone();
 
-/// Arguments that go into `HBASE_OPTS`, so *not* the heap settings (which go into `HBASE_HEAPSIZE`).
-pub fn construct_role_specific_non_heap_jvm_args(
-    hbase_role: &HbaseRole,
-    role: &Role<HbaseConfigFragment, GenericRoleConfig, JavaCommonConfig>,
-    role_group: &str,
-    product_version: &str,
-    kerberos_enabled: bool,
-) -> Result<String, Error> {
-    let mut jvm_args = construct_role_specific_jvm_args(
-        hbase_role,
-        role,
-        role_group,
-        product_version,
-        kerberos_enabled,
-    )?;
     jvm_args.retain(|arg| !is_heap_jvm_argument(arg));
 
     Ok(jvm_args.join(" "))
@@ -102,10 +115,10 @@ pub fn construct_role_specific_non_heap_jvm_args(
 /// Looking at `bin/hbase`, you can actually add the `m` suffix to make the unit more clear, the
 /// script will detect this [here](https://github.com/apache/hbase/blob/777010361abb203b8b17673d84acf4f7f1d0283a/bin/hbase#L165)
 /// and work correctly.
-pub fn construct_hbase_heapsize_env(merged_config: &HbaseConfig) -> Result<String, Error> {
+pub fn construct_hbase_heapsize_env(merged_config: &AnyServiceConfig) -> Result<String, Error> {
     let heap_size = MemoryQuantity::try_from(
         merged_config
-            .resources
+            .resources()
             .memory
             .limit
             .as_ref()
@@ -154,17 +167,15 @@ mod tests {
               default:
                 replicas: 1
         "#;
-        let (hbase_role, merged_config, role, role_group, product_version) =
+        let (hbase, hbase_role, merged_config, role_group, product_version) =
             construct_boilerplate(input);
-        let kerberos_enabled = false;
 
-        let global_jvm_args = construct_global_jvm_args(kerberos_enabled);
+        let global_jvm_args = construct_global_jvm_args(false);
         let role_specific_non_heap_jvm_args = construct_role_specific_non_heap_jvm_args(
+            &hbase,
             &hbase_role,
-            &role,
             &role_group,
             &product_version,
-            kerberos_enabled,
         )
         .unwrap();
         let hbase_heapsize_env = construct_hbase_heapsize_env(&merged_config).unwrap();
@@ -190,6 +201,10 @@ mod tests {
           clusterConfig:
             hdfsConfigMapName: simple-hdfs
             zookeeperConfigMapName: simple-znode
+            authentication:
+              tlsSecretClass: tls
+              kerberos:
+                secretClass: kerberos-simple
           masters:
             roleGroups:
               default:
@@ -214,17 +229,15 @@ mod tests {
                     - -Xmx40000m # This has no effect!
                     - -Dhttps.proxyPort=1234
         "#;
-        let (hbase_role, merged_config, role, role_group, product_version) =
+        let (hbase, hbase_role, merged_config, role_group, product_version) =
             construct_boilerplate(input);
-        let kerberos_enabled = true;
 
-        let global_jvm_args = construct_global_jvm_args(kerberos_enabled);
+        let global_jvm_args = construct_global_jvm_args(hbase.has_kerberos_enabled());
         let role_specific_non_heap_jvm_args = construct_role_specific_non_heap_jvm_args(
+            &hbase,
             &hbase_role,
-            &role,
             &role_group,
             &product_version,
-            kerberos_enabled,
         )
         .unwrap();
         let hbase_heapsize_env = construct_hbase_heapsize_env(&merged_config).unwrap();
@@ -247,27 +260,19 @@ mod tests {
 
     fn construct_boilerplate(
         hbase_cluster: &str,
-    ) -> (
-        HbaseRole,
-        HbaseConfig,
-        Role<HbaseConfigFragment, GenericRoleConfig, JavaCommonConfig>,
-        String,
-        String,
-    ) {
+    ) -> (HbaseCluster, HbaseRole, AnyServiceConfig, String, String) {
         let hbase: HbaseCluster = serde_yaml::from_str(hbase_cluster).expect("illegal test input");
 
         let hbase_role = HbaseRole::RegionServer;
         let merged_config = hbase
             .merged_config(&hbase_role, "default", "my-hdfs")
             .unwrap();
-        let role: Role<HbaseConfigFragment, GenericRoleConfig, JavaCommonConfig> =
-            hbase.spec.region_servers.unwrap();
         let product_version = hbase.spec.image.product_version().to_owned();
 
         (
+            hbase,
             hbase_role,
             merged_config,
-            role,
             "default".to_owned(),
             product_version,
         )