+++ CHANGELOG.md
@@ -26,12 +26,17 @@
 
 ## [24.11.1] - 2025-01-09
 
+### Changed
+
+- Support moving regions to other Pods during graceful shutdown of region servers ([#570]).
+
 ### Fixed
 
 - BREAKING: Use distinct ServiceAccounts for the Stacklets, so that multiple Stacklets can be
   deployed in one namespace. Existing Stacklets will use the newly created ServiceAccounts after
   restart ([#594]).
 
+[#570]: https://github.com/stackabletech/hbase-operator/pull/570
 [#594]: https://github.com/stackabletech/hbase-operator/pull/594
 
 ## [24.11.0] - 2024-11-18